+++ src/main/java/uk/gov/hmcts/reform/blobrouter/clients/bulkscanprocessor/BulkScanProcessorClient.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import uk.gov.hmcts.reform.blobrouter.clients.model.response.SasTokenResponse;
+import uk.gov.hmcts.reform.blobrouter.clients.response.SasTokenResponse;
 import uk.gov.hmcts.reform.blobrouter.reconciliation.report.ReconciliationReportResponse;
 import uk.gov.hmcts.reform.blobrouter.reconciliation.report.ReconciliationStatement;
