+++ src/main/java/uk/gov/hmcts/reform/blobrouter/clients/response/SasTokenResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.blobrouter.clients.model.response;
+package uk.gov.hmcts.reform.blobrouter.clients.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
