+++ src/test/java/uk/gov/hmcts/reform/blobrouter/services/storage/BulkScanSasTokenCacheTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.blobrouter.clients.bulkscanprocessor.BulkScanProcessorClient;
-import uk.gov.hmcts.reform.blobrouter.clients.model.response.SasTokenResponse;
+import uk.gov.hmcts.reform.blobrouter.clients.response.SasTokenResponse;
 import uk.gov.hmcts.reform.blobrouter.exceptions.InvalidSasTokenException;
 
 import java.time.OffsetDateTime;
